move adapters into data subfolders
