+++ __tests__/controllers/cacheController.spec.ts
@@ -1,13 +1,13 @@
 import { beforeAll, beforeEach, describe, expect, it, jest } from '@jest/globals'
 import { NotionMock } from '../support/notionMock'
 import { getMockReq, getMockRes } from '@jest-mock/express'
-import NotionAdapter from '../../src/data/notionAdapter'
+import NotionAdapter from '../../src/data/notion/notionAdapter'
 import CacheController from '../../src/controllers/cacheController'
 import { transaction } from '../../__mocks__/firebase/firestore'
 import { Request } from 'express'
 import Week from '../../src/models/week'
 import { FirebaseMock } from '../support/firebaseMock'
-import FirestoreAdapter from '../../src/data/firestoreAdapter'
+import FirestoreAdapter from '../../src/data/firestore/firestoreAdapter'
 import { NotionMovie } from '../support/notionHelpers'
 import { TmdbMock } from '../support/tmdbMock'
 import { mockFetch } from '../support/fetchMock'
