+++ src/data/notion/notionAdapter.ts
@@ -1,13 +1,13 @@
 import { Client, isFullPageOrDatabase } from '@notionhq/client'
-import Movie from '../models/movie.js'
-import Week from '../models/week.js'
+import Movie from '../../models/movie.js'
+import Week from '../../models/week.js'
 import {
   PartialDatabaseObjectResponse,
   type PageObjectResponse,
   type PartialPageObjectResponse,
   DatabaseObjectResponse,
 } from '@notionhq/client/build/src/api-endpoints'
-import type WeekProperties from '../types/weekProperties.js'
+import type WeekProperties from '../../types/weekProperties.js'
 
 export default class NotionAdapter {
   #notion: Client
