+++ src/index.ts
@@ -1,7 +1,7 @@
 import * as dotenv from 'dotenv'
 import Application from './application.js'
-import NotionAdapter from './data/notionAdapter.js'
-import FirestoreAdapter from './data/firestoreAdapter.js'
+import NotionAdapter from './data/notion/notionAdapter.js'
+import FirestoreAdapter from './data/firestore/firestoreAdapter.js'
 import setupExpress from './config/express.js'
 import TmdbAdapter from './data/tmdb/tmdbAdapter.js'
